feat(render): activate Rust layer iteration

+++ core/fpdfapi/render/cpdf_rendercontext.h
@@ -64,7 +64,7 @@ class CPDF_RenderContext {
 #endif
 
   size_t CountLayers() const { return layers_.size(); }
-  Layer* GetLayer(uint32_t index) { return &layers_[index]; }
+  Layer* GetLayer(size_t index) { return &layers_[index]; }
 
   CPDF_Document* GetDocument() const { return document_; }
   const CPDF_Dictionary* GetPageResources() const {
